Added 3 new goals - increment, add-snapshot, remove-snapshot
